added authentication to task routes;

+++ src/routes/taskRoutes.js
@@ -1,14 +1,21 @@
 const express = require('express') // Require Express Module
 const Task = require('../models/task') // Bring over the model for this resource
+const auth = require('../middleware/auth')
 
 const router = new express.Router() // Create the new Router instance
 
 
 // 1. Create Task - Async/Await Style
 
-router.post('/tasks', async (req,res) => {
+router.post('/tasks', auth, async (req,res) => {
     
-    const task = new Task(req.body) 
+    // const task = new Task(req.body) 
+
+    const task = new Task({
+        ...req.body, // ES6 method for copying all properties from the request body to this object
+        owner: req.user._id
+    })
+
 
     try {
         const result = await task.save()
@@ -21,11 +28,12 @@ router.post('/tasks', async (req,res) => {
 
 // 2. Get All Tasks - Async/Await Style
 
-router.get('/tasks', async (req,res) => {
+router.get('/tasks', auth, async (req,res) => {
 
     try {
-        const tasks = await Task.find({})
-        res.status(200).send(tasks)
+        // const tasks = await Task.find({ owner: req.user._id })
+        await req.user.populate('tasks').execPopulate()
+        res.send(req.user.tasks)
 
     } catch (e) {
         res.status(500).send()
@@ -34,12 +42,13 @@ router.get('/tasks', async (req,res) => {
 
 // 3. Get Task By ID - Async/Await Style
 
-router.get('/tasks/:id', async (req,res) => {
+router.get('/tasks/:id', auth, async (req,res) => {
 
     const _id = req.params.id
 
     try {
-        const task = await Task.findById(_id)
+        // const task = await Task.findById(_id)
+        const task = await Task.findOne({ _id, owner: req.user_id })
 
         if (!task) {
             res.status(404).send()
@@ -54,28 +63,31 @@ router.get('/tasks/:id', async (req,res) => {
 
 // 4. Update Task By ID - Async/Await Style
 
-router.patch('/tasks/:id', async (req,res) => {
+router.patch('/tasks/:id', auth, async (req,res) => {
     
-    const _id = req.params.id // access and save the value passed in
+    // const _id = req.params.id // access and save the value passed in
     const updates = Object.keys(req.body) // stores all of the keys passed in req.body in a new array object
-    const allowedUpdates = ["taskDescr, isCompleted"] // array that defines what keys we will allowed to be updated
-    
-    // every() takes a callback as it's only argument
-    // for every item in the updates array check to see if that value is included in the allowedUpdates array
-    // returns a boolean (true or false)
-    const isValid = updates.every((update) => allowedUpdates.includes(update)) 
+    const allowedUpdates = ['taskDescr', 'isCompleted'] // array that defines what keys we will allowed to be updated
+    const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
+ 
 
-    if (!isValid) {
+    if (!isValidOperation) {
         return res.status(400).send({ error: 'Invalid updates!' })
     }
 
     try {
-        const task = await Task.findOneAndUpdate(_id, req.body, { new: true, runValidators: true }) 
+        // const task = await Task.findOneAndUpdate(_id, req.body, { new: true, runValidators: true }) 
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id }) 
         
         if (!task) {
             return res.status(404).send()
         }
 
+        // every() takes a callback as it's only argument
+        // for every item in the updates array check to see if that value is included in the allowedUpdates array
+        // returns a boolean (true or false)
+        updates.forEach((update) => task[update] = req.body[update]) 
+        await task.save()
         res.send(task)
     } catch (e) {
         res.status(500).send(e)
@@ -84,12 +96,12 @@ router.patch('/tasks/:id', async (req,res) => {
 
 // 5. Delete Task By ID - Async/Await Style
 
-router.delete('/tasks/:id', async (req,res) => {
+router.delete('/tasks/:id', auth, async (req,res) => {
 
-    const _id = req.params.id
+   // const _id = req.params.id
 
     try {
-        const task = await Task.findByIdAndDelete(_id)
+        const task = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id })
 
     if(!task) {
         res.status(404).send()
